django-boilerplate/requirements.txt
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-appconf==0.6
django-bootstrap3==5.0.3
django-ckeditor==4.4.6
django-compressor==1.4
django-extensions==1.3.6
django-grappelli==2.6.1
django-libsass==0.2
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==3.0.2
drf-nested-routers==0.9.0
facebook-sdk==0.4.0
gunicorn==18.0
libsass==0.7.0
Pillow==2.4.0
pkg-resources==0.0.0
psycopg2==2.5.2
pystache==0.5.4
requests==2.2.1
six==1.6.1
static==1.0.2
static3==0.5.1
######
